x11/cliprdr: correctly trim terminating null bytes from strings
Sometimes Windows sends strings with excess null terminating bytes.
For example, when one copies digits from calc.exe (issue #2287).

All trailing null bytes should be trimmed from strings, not only the
last one. Because, for example, LibreOffice freaks out if these are
present and replaces them with # characters.

Also, if the trailing null byte is not present, the length should not
be adjusted. For example, Firefox actually sends "HTML Format" without
one, thus we mistakingly deleted the '>' character of "</html>" tag.
This is not the case now as we do check for the null byte.