x11/cliprdr: create temporary directory for file transfers
We will need a temporary directory to store files downloaded from
the remote clipboard. The directory will be used session-wide so
it is initialized on cliprdr startup and removed during cleanup.

This requires X/Open 7 conformance (equivalent to POSIX.1-2008)
to get access to functions mkdtemp() and nftw(). They are just
too convenient to use, sorry.

Here are some tips to relieve the anguish of those who encounter
compilation errors in their old environments:

  - An alternative for nftw() would be to manually implement DFS
    with readdir(). Not that hard, but clunky if done correctly.

  - mkdtemp() does not have a safe alternative in older standards.
    Just be creative and enjoy your possible race conditions.