x11/cliprdr: negotiate file streaming support
The spec (MS-RDPECLIP 3.2.5.1.3) says that we should not ask for file
clipping support if it was not advertised by the server side. Therefore
we'd better check for this capability flag.

There is no command-line switch to disable file clipping, so we always
support it. Thus, file clipping support can be enabled and advertised
in client capabilities if the server supports it and the client has
successfully created a temporary directory required for file clipping.

We also set the CB_FILECLIP_NO_FILE_PATHS flag in capabilities because
it seems to be necessary for the server to send "FileGroupDescriptorW"
format to us. Otherwise the server will only send old CF_HDROP format
which can't be handled well without enabled disk drive redirection and
properly negotiated server-side temporary directory.