wClipboard/fuse: create and cleanup mount point
FUSE will need a directory to mount to so we should create one. We also
should remove that directory when we're done and don't need it anymore.

First try to create a directory in /var/run/user hierarchy available on
many Linux systems, and if that fails then fall back to using /tmp.

When creating a directory make sure that all parent directories are
present as well. Also don't fail if the directory is already there.

Make sure that different FreeRDP processes launched by different users
get different directories which can be used independently.