wClipboard/fuse: register FileGroupDescritorW synthesizer
This starts work on the file list handling. In order to get the files we need
to register a synthesizer providing conversion from FileGroupDescriptorW into
text/uri-list (requested by xfreerdp).

The synthesizer first performs some basic sanity checks to ensure that it is
called with the right input data (an array of FILEDESCRIPTOR structs). Then it
converts these common WinPR structs into fuse_file structs which are stored in
subsystem->remote_files list for later processing and future reference. After
that the fuse_file structs need to be converted into actual text/uri-list
string expected by the client.

Preprocessing the files and converting the file list will be covered in later
commits.