wClipboard/fuse: convert FILEDESCRIPTORs to fuse_files
Now we can start working on the conversion. The first step is to convert
an array of FILEDESCRIPTOR structs into a flat array of fuse_file structs.
This is more or less straightforward conversion, with filename processing
being the only interestring point.

Why a separate array instead of directly populating ArrayList? Well, because
FUSE needs to know the exact hierarchy of the files and directories which is
not apparent in the flat file list. We are going to reconstruct the directory
structure in the next commit, so keeping the files in a separate container is
more convenient. Stay tuned.