wClipboard/fuse: reconstruct directory structure
Now we have a flat file list, but we actually need a tree. The 'contents' lists
in our files are currently empty. In order to populate them we need to know
which files go into which directories. This information can only be derived
from the file names: the files with the same prefixes are located in the same
directory. This invariant can be used to reconstruct the directory structure.

In order to efficiently locate prefixes we'll sort the file list. This places
files which are in the same directory next to each other in the array. After
that a linear pass through the array can relocate all the files into their
correct places.

We use the qsort() function from the standard library because it's good enough
and we don't have to write our own sort. Usually the file list from the server
is already sorted or almost sorted. Hopefully the library implementation takes
that into account and does a pre-sort check for that.

After the files are sorted we can reorder them. A simple recursive algoritm is
used for this. The push_files_rec() function keeps pushing consecutive files
into the first array entry (which should be a directory) until the prefix stops
matching. It's somewhat inefficient in that it could scan for the first not
matching prefix beforehand, but that implementation is much more complex.
There are some tracing logs in the function in order to help debugging it
if^W when it breaks. Hopefully it doesn't.