wClipboard/fuse: stat() and basic file lookup
Implement getattr() callback which is backing up stat() calls in FUSE.
This requires us to touch the 'easy' FUSE API which is based on file
paths. We need to be able to find files by paths. Preferably, fast, but
we are limited by FUSE API here. We will cache the files if possible.

lookup_fuse_file() is the main function which finds files by paths. It
converts a FUSE path (e.g., "/file.txt" for file named "file.txt") into
corresponding struct fuse_file*, or returns an error code.

Lookup involves some trickery with slashes, but otherwise it is a
transparent, component-by-component downward movement through
subdirectories until we exhaust the path. If there is no component then
the file is missing. There is another error if a component is not a
directory while the path insists it should be one. Though, these should
never happen in practice.

Note that we also have to fill in the statbuf structure. We fill in
only bare minimum: inode number and file mode to differentiate between
files and directories. We'll get to the size later and we don't really
care about other fields.