wClipboard/fuse: directory traversal
Implement readdir() and company. This allows the system to actually
paste the files somewhere, preserving the directory structure. The files
themselves are empty though.

Here we're introducing filename lookup caching: opendir() looks up the
directory which is then reused in readdir(). This is not strictly
necessary as our readdir() implementation does all job in one sweep, but
it may become useful later.