wClipboard: use x-special/gnome-copied-files for file transfers
It turned out that Nautilus does not request text/uri-list when asked to
paste files. (Thunar supports both.) While text/uri-list seems to be
less specific format, not locked to GNOME and Linux, there is no use for
a format which does not work.

Currently I do not have a specific solution to this problem in my mind.
I guess we'll have to add more conversions to WinPR's implementation and
update the clients accordingly to use the new formats.

For now let's use x-special/gnome-copied-files format which is supported
by Nautilus. Update all of the clipboard code to support it: the X11
client as well as POSIX and FUSE subsystems.