wClipboard/fuse: fix FUSE subsystem termination
Unfortunately, FUSE subsystems blocks the execution when it is
terminated. It gets stuck in fuse_loop(), regardless of our fuse_exit()
calls. They only make FUSE exit when it is out of some inner loop which
happens only when some filesystem request arrives. I've been doing
"ls /var/run/user/$PID/winprd/clipboard-*" to fix this, but that's dumb.

It seems that the only real way to get this fixes is to use POSIX
signals. They are able to interrupt the read() system call where
fuse_loop() is blocked and then it actually terminates. However, signals
are so inconvenient in threaded environment that I'd rather use
pthread_cancel(). (And yes, using fuse_loop_mt() does not help.)